details in description below
last commit on this repo

Why? Dis-Snek was rebranded to N.A.F.F. and had many breaking changes; it's easier to just remake everything instead of trying to fix it.
I'll keep basics and the V3 will look very similar to this repo.

See you in https://github.com/Alberto-X3/AlbertoX3-V3